Move compilation to abstract test
